scheduler/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Flask==2.2.2
Flask-JWT-Extended==4.4.4
flask-serialize==2.1.1
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
permissive-dict==1.0.4
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
PyJWT==2.5.0
pyparsing==3.0.9
pytz==2022.4
redis==4.3.4
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.42
SQLAlchemy-serializer==1.4.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
google-cloud-storage==2.6.0
